ai-agent/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
psycopg2-binary==2.9.11
SQLAlchemy==2.0.49
pandas==2.2.1
python-dotenv==1.2.1
google-genai==0.3.0
requests==2.31.0
beautifulsoup4==4.12.3
googlesearch-python==1.2.5
# Multi-provider AI fallback (opcionales según cuáles uses)
openai>=1.0.0
anthropic>=0.25.0
